GitHub Actions: version artifact name
merges https://github.com/manubot/rootstock/pull/305

Create artifacts that are named like manuscript-32632624-d1c4caf
and contain the GitHub Actions run-id as well as the head of the triggering
commit hash. Note that on a pull request, this is the commit that was
added to the PR and not the auto-generated merge commit that the
action is actually evaluating.